+++ ApexGameEngine/src/Platform/OpenGL/OpenGLBuffer.h
@@ -26,7 +26,7 @@ namespace Apex {
 	class OpenGLIndexBuffer : public IndexBuffer
 	{
 	public:
-		OpenGLIndexBuffer(uint32_t* indices, size_t count);
+		OpenGLIndexBuffer(uint32_t* indices, uint32_t count);
 		virtual ~OpenGLIndexBuffer();
 
 		virtual void Bind() const override;
